requirements.txt
solana==0.30.2
solders>=0.18.0
anchorpy>=0.17.0
websockets>=11.0.3
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.5.0
aiohttp>=3.8.5
pydantic>=2.4.0
flask>=2.3.3
flask-cors>=4.0.0 